+++ Countly.js
@@ -7,6 +7,8 @@
 import { Platform, NativeModules, NativeEventEmitter } from 'react-native';
 import parseErrorStackLib from '../react-native/Libraries/Core/Devtools/parseErrorStack.js';
 import CountlyConfig from './CountlyConfig.js';
+import CountlyState from './CountlyState.js';
+import Feedback from './Feedback.js';
 
 const { CountlyReactNative } = NativeModules;
 const eventEmitter = new NativeEventEmitter(CountlyReactNative);
@@ -14,7 +16,13 @@ const eventEmitter = new NativeEventEmitter(CountlyReactNative);
 const Countly = {};
 Countly.serverUrl = '';
 Countly.appKey = '';
-let _isInitialized = false;
+let _state = CountlyState;
+CountlyState.CountlyReactNative = CountlyReactNative;
+
+Countly.feedback = Feedback;
+Countly.feedback.state = CountlyState;
+Countly.feedback.instance = Countly;
+
 let _isPushInitialized = false;
 const DeviceIdType = {
     DEVELOPER_SUPPLIED: 'DEVELOPER_SUPPLIED',
@@ -50,7 +58,7 @@ Countly.TemporaryDeviceIDString = 'TemporaryDeviceID';
 /**
  * Initialize Countly
  *
- * @deprecated in 23.02.0 : use 'initWithConfig' intead of 'init'.
+ * @deprecated in 23.02.0 : use 'initWithConfig' instead of 'init'.
  *
  * @function Countly.init should be used to initialize countly
  * @param {String} serverURL server url
@@ -70,7 +78,7 @@ Countly.init = async function (serverUrl, appKey, deviceId) {
  * @param {Object} countlyConfig countly config object
  */
 Countly.initWithConfig = async function (countlyConfig) {
-    if (_isInitialized) {
+    if (_state.isInitialized) {
         Countly.logError('init', 'SDK is already initialized');
         return;
     }
@@ -92,7 +100,7 @@ Countly.initWithConfig = async function (countlyConfig) {
     const argsString = JSON.stringify(argsMap);
     args.push(argsString);
     await CountlyReactNative.init(args);
-    _isInitialized = true;
+    _state.isInitialized = true;
 };
 
 /**
@@ -189,8 +197,8 @@ _configToJson = function (config) {
  */
 Countly.isInitialized = async function () {
     // returns a promise
-    _isInitialized = await CountlyReactNative.isInitialized();
-    return _isInitialized;
+    _state.isInitialized = await CountlyReactNative.isInitialized();
+    return _state.isInitialized;
 };
 
 /**
@@ -203,7 +211,7 @@ Countly.isInitialized = async function () {
  * if true, countly sdk has started
  */
 Countly.hasBeenCalledOnStart = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'hasBeenCalledOnStart'";
         Countly.logError('hasBeenCalledOnStart', message);
         return message;
@@ -219,7 +227,7 @@ Countly.hasBeenCalledOnStart = function () {
  * @return {String || void} error message or void
  */
 Countly.sendEvent = function (options) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'sendEvent'";
         Countly.logError('sendEvent', message);
         return message;
@@ -278,9 +286,7 @@ Countly.sendEvent = function (options) {
  *
  */
 Countly.setViewTracking = async function (boolean) {
-    if (await CountlyReactNative.isLoggingEnabled()) {
-        console.log('[CountlyReactNative] setViewTracking is deprecated.');
-    }
+    Countly.logWarning('setViewTracking', 'setViewTracking is deprecated.');
 };
 
 /**
@@ -292,7 +298,7 @@ Countly.setViewTracking = async function (boolean) {
  * @return {String || void} error message or void
  */
 Countly.recordView = async function (recordView, segments) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'recordView'";
         Countly.logError('recordView', msg);
         return msg;
@@ -329,7 +335,7 @@ Countly.disablePushNotifications = function () {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.pushTokenType' intead of 'pushTokenType'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.pushTokenType' instead of 'pushTokenType'.
  *
  * Set messaging mode for push notifications
  * Should be called before Countly init
@@ -364,7 +370,7 @@ Countly.sendPushToken = function (options) {
  *
  */
 Countly.askForNotificationPermission = function (customSoundPath = 'null') {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'askForNotificationPermission'";
         Countly.logError('askForNotificationPermission', message);
         return message;
@@ -386,7 +392,7 @@ Countly.registerForNotification = function (theListener) {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.configureIntentRedirectionCheck' intead of 'configureIntentRedirectionCheck'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.configureIntentRedirectionCheck' instead of 'configureIntentRedirectionCheck'.
  *
  * Configure intent redirection checks for push notification
  * Should be called before Countly "askForNotificationPermission"
@@ -496,7 +502,7 @@ Countly.setLoggingEnabled = function (enabled = true) {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.setLocation' intead of 'setLocationInit'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.setLocation' instead of 'setLocationInit'.
  *
  * Set user initial location
  * Should be called before init
@@ -523,7 +529,7 @@ Countly.setLocationInit = function (countryCode, city, location, ipAddress) {
  * @param {IP address of user's} ipAddress
  * */
 Countly.setLocation = function (countryCode, city, location, ipAddress) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'setLocation'";
         Countly.logError('setLocation', message);
         return message;
@@ -543,7 +549,7 @@ Countly.setLocation = function (countryCode, city, location, ipAddress) {
  * @return {String || void} error message or void
  */
 Countly.disableLocation = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'disableLocation'";
         Countly.logError('disableLocation', message);
         return message;
@@ -559,7 +565,7 @@ Countly.disableLocation = function () {
  * @return {String} device id or error message
  */
 Countly.getCurrentDeviceId = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'getCurrentDeviceId'";
         Countly.logError('getCurrentDeviceId', message);
         return message;
@@ -601,7 +607,7 @@ _getDeviceIdType = function (deviceIdType) {
  * @return {DeviceIdType || null} deviceIdType or null
  * */
 Countly.getDeviceIDType = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         Countly.logError('getDeviceIDType', "'init' must be called before 'getDeviceIDType'");
         return null;
     }
@@ -621,7 +627,7 @@ Countly.getDeviceIDType = async function () {
  * @return {String || void} error message or void
  * */
 Countly.changeDeviceId = async function (newDeviceID, onServer) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'changeDeviceId'";
         Countly.logError('changeDeviceId', msg);
         return msg;
@@ -654,7 +660,7 @@ Countly.setHttpPostForced = function (boolean = true) {
 
 Countly.isCrashReportingEnabled = false;
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.enableCrashReporting' intead of 'enableCrashReporting'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.enableCrashReporting' instead of 'enableCrashReporting'.
  *
  * Enable crash reporting to report unhandled crashes to Countly
  * Should be called before Countly init
@@ -662,9 +668,7 @@ Countly.isCrashReportingEnabled = false;
 Countly.enableCrashReporting = async function () {
     CountlyReactNative.enableCrashReporting();
     if (ErrorUtils && !Countly.isCrashReportingEnabled) {
-        if (await CountlyReactNative.isLoggingEnabled()) {
-            console.log('[CountlyReactNative] Adding Countly JS error handler.');
-        }
+        Countly.logInfo('enableCrashReporting', 'Adding Countly JS error handler.');
         const previousHandler = ErrorUtils.getGlobalHandler();
         ErrorUtils.setGlobalHandler((error, isFatal) => {
             const jsStackTrace = Countly.getStackTrace(error);
@@ -729,7 +733,7 @@ Countly.getStackTrace = (e) => {
             jsStackTrace = parseErrorStackLib(e);
         }
     } catch (e) {
-        // console.log(e.message);
+        // Countly.logError('getStackTrace', e.message);
     }
     return jsStackTrace;
 };
@@ -742,7 +746,7 @@ Countly.getStackTrace = (e) => {
  * @return {String || void} error message or void
  */
 Countly.addCrashLog = function (crashLog) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'addCrashLog'";
         Countly.logError('addCrashLog', message);
         return message;
@@ -760,7 +764,7 @@ Countly.addCrashLog = function (crashLog) {
  * @return {String || void} error message or void
  */
 Countly.logException = function (exception, nonfatal, segments) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'logException'";
         Countly.logError('logException', message);
         return message;
@@ -802,7 +806,7 @@ Countly.setCustomCrashSegments = function (segments) {
  * @return {String || void} error message or void
  */
 Countly.startSession = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'startSession'";
         Countly.logError('startSession', message);
         return message;
@@ -817,7 +821,7 @@ Countly.startSession = function () {
  * @return {String || void} error message or void
  */
 Countly.endSession = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'endSession'";
         Countly.logError('endSession', message);
         return message;
@@ -826,7 +830,7 @@ Countly.endSession = function () {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.enableParameterTamperingProtection' intead of 'enableParameterTamperingProtection'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.enableParameterTamperingProtection' instead of 'enableParameterTamperingProtection'.
  *
  * Set the optional salt to be used for calculating the checksum of requested data which will be sent with each request, using the &checksum field
  * Should be called before Countly init
@@ -867,7 +871,7 @@ Countly.pinnedCertificates = async function (certificateName) {
  * @return {String || void} error message or void
  */
 Countly.startEvent = async function (eventName) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'startEvent'";
         Countly.logError('startEvent', msg);
         return msg;
@@ -888,7 +892,7 @@ Countly.startEvent = async function (eventName) {
  * @return {String || void} error message or void
  */
 Countly.cancelEvent = async function (eventName) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'cancelEvent'";
         Countly.logError('cancelEvent', msg);
         return msg;
@@ -909,7 +913,7 @@ Countly.cancelEvent = async function (eventName) {
  * @return {String || void} error message or void
  */
 Countly.endEvent = function (options) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'endEvent'";
         Countly.logError('endEvent', message);
         return message;
@@ -973,7 +977,7 @@ Countly.endEvent = function (options) {
  * @return {String || void} error message or void
  */
 Countly.setUserData = async function (userData) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setUserData'";
         Countly.logError('setUserData', msg);
         return msg;
@@ -1012,7 +1016,7 @@ Countly.setUserData = async function (userData) {
 
 Countly.userData = {};
 Countly.userData.setProperty = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setProperty'";
         Countly.logError('setProperty', msg);
         return msg;
@@ -1033,7 +1037,7 @@ Countly.userData.setProperty = async function (keyName, keyValue) {
     }
 };
 Countly.userData.increment = async function (keyName) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'increment'";
         Countly.logError('increment', msg);
         return msg;
@@ -1048,7 +1052,7 @@ Countly.userData.increment = async function (keyName) {
     }
 };
 Countly.userData.incrementBy = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'incrementBy'";
         Countly.logError('incrementBy', msg);
         return msg;
@@ -1065,7 +1069,7 @@ Countly.userData.incrementBy = async function (keyName, keyValue) {
     await CountlyReactNative.userData_incrementBy([keyName, intValue]);
 };
 Countly.userData.multiply = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'multiply'";
         Countly.logError('multiply', msg);
         return msg;
@@ -1082,7 +1086,7 @@ Countly.userData.multiply = async function (keyName, keyValue) {
     await CountlyReactNative.userData_multiply([keyName, intValue]);
 };
 Countly.userData.saveMax = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'saveMax'";
         Countly.logError('saveMax', msg);
         return msg;
@@ -1099,7 +1103,7 @@ Countly.userData.saveMax = async function (keyName, keyValue) {
     await CountlyReactNative.userData_saveMax([keyName, intValue]);
 };
 Countly.userData.saveMin = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'saveMin'";
         Countly.logError('saveMin', msg);
         return msg;
@@ -1116,7 +1120,7 @@ Countly.userData.saveMin = async function (keyName, keyValue) {
     await CountlyReactNative.userData_saveMin([keyName, intValue]);
 };
 Countly.userData.setOnce = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setOnce'";
         Countly.logError('setOnce', msg);
         return msg;
@@ -1135,7 +1139,7 @@ Countly.userData.setOnce = async function (keyName, keyValue) {
     }
 };
 Countly.userData.pushUniqueValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pushUniqueValue'";
         Countly.logError('pushUniqueValue', msg);
         return msg;
@@ -1154,7 +1158,7 @@ Countly.userData.pushUniqueValue = async function (keyName, keyValue) {
     }
 };
 Countly.userData.pushValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pushValue'";
         Countly.logError('pushValue', msg);
         return msg;
@@ -1173,7 +1177,7 @@ Countly.userData.pushValue = async function (keyName, keyValue) {
     }
 };
 Countly.userData.pullValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pullValue'";
         Countly.logError('pullValue', msg);
         return msg;
@@ -1197,7 +1201,7 @@ Countly.userDataBulk = {};
 
 // providing key/values with predefined and custom properties
 Countly.userDataBulk.setUserProperties = async function (customAndPredefined) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setUserProperties'";
         Countly.logError('setUserProperties', msg);
         return msg;
@@ -1233,7 +1237,7 @@ Countly.userDataBulk.setUserProperties = async function (customAndPredefined) {
 };
 
 Countly.userDataBulk.save = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'save'";
         Countly.logError('save', msg);
         return msg;
@@ -1242,7 +1246,7 @@ Countly.userDataBulk.save = async function () {
 };
 
 Countly.userDataBulk.setProperty = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setProperty'";
         Countly.logError('setProperty', msg);
         return msg;
@@ -1263,7 +1267,7 @@ Countly.userDataBulk.setProperty = async function (keyName, keyValue) {
     }
 };
 Countly.userDataBulk.increment = async function (keyName) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'increment'";
         Countly.logError('increment', msg);
         return msg;
@@ -1278,7 +1282,7 @@ Countly.userDataBulk.increment = async function (keyName) {
     }
 };
 Countly.userDataBulk.incrementBy = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'incrementBy'";
         Countly.logError('incrementBy', msg);
         return msg;
@@ -1295,7 +1299,7 @@ Countly.userDataBulk.incrementBy = async function (keyName, keyValue) {
     await CountlyReactNative.userDataBulk_incrementBy([keyName, intValue]);
 };
 Countly.userDataBulk.multiply = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'multiply'";
         Countly.logError('multiply', msg);
         return msg;
@@ -1312,7 +1316,7 @@ Countly.userDataBulk.multiply = async function (keyName, keyValue) {
     await CountlyReactNative.userDataBulk_multiply([keyName, intValue]);
 };
 Countly.userDataBulk.saveMax = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'saveMax'";
         Countly.logError('saveMax', msg);
         return msg;
@@ -1329,7 +1333,7 @@ Countly.userDataBulk.saveMax = async function (keyName, keyValue) {
     await CountlyReactNative.userDataBulk_saveMax([keyName, intValue]);
 };
 Countly.userDataBulk.saveMin = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'saveMin'";
         Countly.logError('saveMin', msg);
         return msg;
@@ -1346,7 +1350,7 @@ Countly.userDataBulk.saveMin = async function (keyName, keyValue) {
     await CountlyReactNative.userDataBulk_saveMin([keyName, intValue]);
 };
 Countly.userDataBulk.setOnce = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'setOnce'";
         Countly.logError('setOnce', msg);
         return msg;
@@ -1365,7 +1369,7 @@ Countly.userDataBulk.setOnce = async function (keyName, keyValue) {
     }
 };
 Countly.userDataBulk.pushUniqueValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pushUniqueValue'";
         Countly.logError('pushUniqueValue', msg);
         return msg;
@@ -1384,7 +1388,7 @@ Countly.userDataBulk.pushUniqueValue = async function (keyName, keyValue) {
     }
 };
 Countly.userDataBulk.pushValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pushValue'";
         Countly.logError('pushValue', msg);
         return msg;
@@ -1403,7 +1407,7 @@ Countly.userDataBulk.pushValue = async function (keyName, keyValue) {
     }
 };
 Countly.userDataBulk.pullValue = async function (keyName, keyValue) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'pullValue'";
         Countly.logError('pullValue', msg);
         return msg;
@@ -1423,7 +1427,7 @@ Countly.userDataBulk.pullValue = async function (keyName, keyValue) {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.setRequiresConsent' intead of 'setRequiresConsent'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.setRequiresConsent' instead of 'setRequiresConsent'.
  *
  * Set that consent should be required for features to work.
  * Should be called before Countly init
@@ -1443,7 +1447,7 @@ Countly.setRequiresConsent = function (flag) {
  * @return {String || void} error message or void
  */
 Countly.giveConsent = function (args) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'giveConsent'";
         Countly.logError('giveConsent', message);
         return message;
@@ -1458,7 +1462,7 @@ Countly.giveConsent = function (args) {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.giveConsent' intead of 'giveConsentInit'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.giveConsent' instead of 'giveConsentInit'.
  *
  * Give consent for specific features before init.
  * Should be called after Countly init
@@ -1486,7 +1490,7 @@ Countly.giveConsentInit = async function (args) {
  * @return {String || void} error message or void
  */
 Countly.removeConsent = function (args) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'removeConsent'";
         Countly.logError('removeConsent', message);
         return message;
@@ -1508,7 +1512,7 @@ Countly.removeConsent = function (args) {
  * @return {String || void} error message or void
  */
 Countly.giveAllConsent = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'giveAllConsent'";
         Countly.logError('giveAllConsent', message);
         return message;
@@ -1524,7 +1528,7 @@ Countly.giveAllConsent = function () {
  * @return {String || void} error message or void
  */
 Countly.removeAllConsent = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'removeAllConsent'";
         Countly.logError('removeAllConsent', message);
         return message;
@@ -1533,7 +1537,7 @@ Countly.removeAllConsent = function () {
 };
 
 Countly.remoteConfigUpdate = function (callback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'remoteConfigUpdate'";
         Countly.logError('remoteConfigUpdate', message);
         callback(message);
@@ -1545,7 +1549,7 @@ Countly.remoteConfigUpdate = function (callback) {
 };
 
 Countly.updateRemoteConfigForKeysOnly = function (keyNames, callback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'updateRemoteConfigForKeysOnly'";
         Countly.logError('updateRemoteConfigForKeysOnly', message);
         callback(message);
@@ -1563,7 +1567,7 @@ Countly.updateRemoteConfigForKeysOnly = function (keyNames, callback) {
 };
 
 Countly.updateRemoteConfigExceptKeys = function (keyNames, callback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'updateRemoteConfigExceptKeys'";
         Countly.logError('updateRemoteConfigExceptKeys', message);
         callback(message);
@@ -1581,7 +1585,7 @@ Countly.updateRemoteConfigExceptKeys = function (keyNames, callback) {
 };
 
 Countly.getRemoteConfigValueForKey = function (keyName, callback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'getRemoteConfigValueForKey'";
         Countly.logError('getRemoteConfigValueForKey', message);
         callback(message);
@@ -1592,7 +1596,7 @@ Countly.getRemoteConfigValueForKey = function (keyName, callback) {
             try {
                 value = JSON.parse(value);
             } catch (e) {
-                // console.log(e.message);
+                // Countly.logError('getRemoteConfigValueForKey', e.message);
                 // noop. value will remain string if not JSON parsable and returned as string
             }
         }
@@ -1601,7 +1605,7 @@ Countly.getRemoteConfigValueForKey = function (keyName, callback) {
 };
 
 Countly.getRemoteConfigValueForKeyP = function (keyName) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'getRemoteConfigValueForKeyP'";
         Countly.logError('getRemoteConfigValueForKeyP', message);
         callback(message);
@@ -1616,7 +1620,7 @@ Countly.getRemoteConfigValueForKeyP = function (keyName) {
             try {
                 value = JSON.parse(value);
             } catch (e) {
-                // console.log(e.message);
+                // Countly.logError('getRemoteConfigValueForKeyP', e.message);
                 // noop. value will remain string if not JSON parsable and returned as string
             }
         }
@@ -1627,7 +1631,7 @@ Countly.getRemoteConfigValueForKeyP = function (keyName) {
 };
 
 Countly.remoteConfigClearValues = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'remoteConfigClearValues'";
         Countly.logError('remoteConfigClearValues', message);
         callback(message);
@@ -1637,7 +1641,7 @@ Countly.remoteConfigClearValues = async function () {
     return result;
 };
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.setStarRatingDialogTexts' intead of 'setStarRatingDialogTexts'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.setStarRatingDialogTexts' instead of 'setStarRatingDialogTexts'.
  *
  * Set's the text's for the different fields in the star rating dialog. Set value null if for some field you want to keep the old value
  *
@@ -1655,7 +1659,7 @@ Countly.setStarRatingDialogTexts = function (starRatingTextTitle, starRatingText
 };
 
 Countly.showStarRating = function (callback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'showStarRating'";
         Countly.logError('showStarRating', message);
         return message;
@@ -1672,10 +1676,10 @@ Countly.showStarRating = function (callback) {
  * @param {String} widgetId - id of rating widget to present
  * @param {String} closeButtonText - text for cancel/close button
  * @return {String || void} error message or void
- * @deprecated use 'presentRatingWidgetWithID' intead of 'showFeedbackPopup'.
+ * @deprecated use 'presentRatingWidgetWithID' instead of 'showFeedbackPopup'.
  */
 Countly.showFeedbackPopup = function (widgetId, closeButtonText) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'showFeedbackPopup'";
         Countly.logError('showFeedbackPopup', message);
         return message;
@@ -1691,7 +1695,7 @@ Countly.showFeedbackPopup = function (widgetId, closeButtonText) {
  * @param {callback listener} ratingWidgetCallback
  */
 Countly.presentRatingWidgetWithID = function (widgetId, closeButtonText, ratingWidgetCallback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         var message = "'init' must be called before 'presentRatingWidgetWithID'";
         Countly.logError('presentRatingWidgetWithID', message);
         return message;
@@ -1717,11 +1721,12 @@ Countly.presentRatingWidgetWithID = function (widgetId, closeButtonText, ratingW
 
 /**
  * Get a list of available feedback widgets as array of object to handle multiple widgets of same type.
+ * @deprecated in 23.8.0 : use 'Countly.feedback.getAvailableFeedbackWidgets' instead of 'getFeedbackWidgets'.
  * @param {callback listener} onFinished - returns (retrievedWidgets, error)
  * @return {String || []} error message or []
  */
 Countly.getFeedbackWidgets = async function (onFinished) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'getFeedbackWidgets'";
         Countly.logError('getFeedbackWidgets', message);
         return message;
@@ -1741,7 +1746,7 @@ Countly.getFeedbackWidgets = async function (onFinished) {
 
 /**
  * Get a list of available feedback widgets for this device ID
- * @deprecated in 20.11.1 : use 'getFeedbackWidgets' intead of 'getAvailableFeedbackWidgets'.
+ * @deprecated in 20.11.1 : use 'Countly.feedback.getAvailableFeedbackWidgets' instead of 'getAvailableFeedbackWidgets'.
  * Using the old function it will not be possible to see all the available feedback widgets.
  * In case there are multiple ones for the same type, only the last one will be returned due to their id being overwritten in the type field.
  * The newer function allow also to see the widgets 'name' field which can be further used to filter and identify specific widgets.
@@ -1749,7 +1754,7 @@ Countly.getFeedbackWidgets = async function (onFinished) {
  * @return {String || void} error message or void
  */
 Countly.getAvailableFeedbackWidgets = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'getAvailableFeedbackWidgets'";
         Countly.logError('getAvailableFeedbackWidgets', message);
         return message;
@@ -1761,6 +1766,7 @@ Countly.getAvailableFeedbackWidgets = async function () {
 /**
  * Present a chosen feedback widget
  *
+ * @deprecated in 23.8.0 : use 'Countly.feedback.presentFeedbackWidget' instead of 'presentFeedbackWidgetObject'.
  * @param {Object} feedbackWidget - feeback Widget with id, type and name
  * @param {String} closeButtonText - text for cancel/close button
  * @param {callback listener} widgetShownCallback - Callback to be executed when feedback widget is displayed
@@ -1769,7 +1775,7 @@ Countly.getAvailableFeedbackWidgets = async function () {
  * @return {String || void} error message or void
  */
 Countly.presentFeedbackWidgetObject = async function (feedbackWidget, closeButtonText, widgetShownCallback, widgetClosedCallback) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const msg = "'init' must be called before 'presentFeedbackWidgetObject'";
         Countly.logError('presentFeedbackWidgetObject', msg);
         return msg;
@@ -1820,10 +1826,10 @@ Countly.presentFeedbackWidgetObject = async function (feedbackWidget, closeButto
  * @param {String} widgetId - id of widget to present
  * @param {String} closeButtonText - text for cancel/close button
  * @return {String || void} error message or void
- * @deprecated in 20.11.1 : use 'presentFeedbackWidgetObject' intead of 'presentFeedbackWidget'.
+ * @deprecated in 20.11.1 : use 'Countly.feedback.presentFeedbackWidget' instead of 'presentFeedbackWidget'.
  */
 Countly.presentFeedbackWidget = function (widgetType, widgetId, closeButtonText) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'presentFeedbackWidget'";
         Countly.logError('presentFeedbackWidget', message);
         return message;
@@ -1845,7 +1851,7 @@ Countly.setEventSendThreshold = function (size) {
 };
 
 Countly.startTrace = function (traceKey) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'startTrace'";
         Countly.logError('startTrace', message);
         return message;
@@ -1856,7 +1862,7 @@ Countly.startTrace = function (traceKey) {
 };
 
 Countly.cancelTrace = function (traceKey) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'cancelTrace'";
         Countly.logError('cancelTrace', message);
         return message;
@@ -1867,7 +1873,7 @@ Countly.cancelTrace = function (traceKey) {
 };
 
 Countly.clearAllTraces = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'clearAllTraces'";
         Countly.logError('clearAllTraces', message);
         return message;
@@ -1877,7 +1883,7 @@ Countly.clearAllTraces = function () {
 };
 
 Countly.endTrace = function (traceKey, customMetric) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'endTrace'";
         Countly.logError('endTrace', message);
         return message;
@@ -1893,7 +1899,7 @@ Countly.endTrace = function (traceKey, customMetric) {
 };
 
 Countly.recordNetworkTrace = function (networkTraceKey, responseCode, requestPayloadSize, responsePayloadSize, startTime, endTime) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'recordNetworkTrace'";
         Countly.logError('recordNetworkTrace', message);
         return message;
@@ -1909,7 +1915,7 @@ Countly.recordNetworkTrace = function (networkTraceKey, responseCode, requestPay
 };
 
 /**
- * @deprecated in 23.02.0 : use 'countlyConfig.enableApm' intead of 'enableApm'.
+ * @deprecated in 23.02.0 : use 'countlyConfig.enableApm' instead of 'enableApm'.
  *
  * Enable APM features, which includes the recording of app start time.
  * Should be called before Countly init
@@ -1920,7 +1926,7 @@ Countly.enableApm = function () {
 };
 
 /**
- * @deprecated in 23.02.0 : use 'Countly.recordIndirectAttribution' intead of 'Countly'.
+ * @deprecated in 23.02.0 : use 'Countly.recordIndirectAttribution' instead of 'Countly'.
  *
  * Enable campaign attribution reporting to Countly.
  * For iOS use "recordAttributionID" instead of "enableAttribution"
@@ -1943,7 +1949,7 @@ Countly.enableAttribution = async function (attributionID = '') {
 
 /**
  *
- * @deprecated in 23.02.0 : use 'Countly.recordIndirectAttribution' intead of 'recordAttributionID'.
+ * @deprecated in 23.02.0 : use 'Countly.recordIndirectAttribution' instead of 'recordAttributionID'.
  *
  * set attribution Id for campaign attribution reporting.
  * Currently implemented for iOS only
@@ -1962,7 +1968,7 @@ Countly.recordAttributionID = function (attributionID) {
  * these requests' app key will be replaced with the current app key.
  */
 Countly.replaceAllAppKeysInQueueWithCurrentAppKey = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'replaceAllAppKeysInQueueWithCurrentAppKey'";
         Countly.logError('replaceAllAppKeysInQueueWithCurrentAppKey', message);
         return message;
@@ -1973,7 +1979,7 @@ Countly.replaceAllAppKeysInQueueWithCurrentAppKey = function () {
  * set direct attribution Id for campaign attribution reporting.
  */
 Countly.recordDirectAttribution = function (campaignType, campaignData) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'recordDirectAttribution'";
         Countly.logError('recordDirectAttribution', message);
         return message;
@@ -1987,7 +1993,7 @@ Countly.recordDirectAttribution = function (campaignType, campaignData) {
  * set indirect attribution Id for campaign attribution reporting.
  */
 Countly.recordIndirectAttribution = function (attributionValues) {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'recordIndirectAttribution'";
         Countly.logError('recordIndirectAttribution', message);
         return message;
@@ -2002,7 +2008,7 @@ Countly.recordIndirectAttribution = function (attributionValues) {
  * these requests will be removed from request queue.
  */
 Countly.removeDifferentAppKeysFromQueue = function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'removeDifferentAppKeysFromQueue'";
         Countly.logError('removeDifferentAppKeysFromQueue', message);
         return message;
@@ -2015,7 +2021,7 @@ Countly.removeDifferentAppKeysFromQueue = function () {
  * Should be called after init.
  */
 Countly.appLoadingFinished = async function () {
-    if (!_isInitialized) {
+    if (!_state.isInitialized) {
         const message = "'init' must be called before 'appLoadingFinished'";
         Countly.logError('appLoadingFinished', message);
         return message;
@@ -2159,6 +2165,28 @@ Countly.validateString = async (stringValue, stringName, functionName) => {
         Countly.logError(functionName, message);
     }
     return message;
+};
+
+/**
+ * Print debug message if logging is enabled
+ * @param {String} functionName : name of function from where value is validating.
+ * @param {String} message : debug message
+ */
+Countly.logDebug = async (functionName, message) => {
+    if (await CountlyReactNative.isLoggingEnabled()) {
+        console.debug(`[CountlyReactNative] ${functionName}, ${message}`);
+    }
+};
+
+/**
+ * Print info messages if logging is enabled
+ * @param {String} functionName : name of function from where value is validating.
+ * @param {String} message : info message
+ */
+Countly.logInfo = async (functionName, message) => {
+    if (await CountlyReactNative.isLoggingEnabled()) {
+        console.info(`[CountlyReactNative] ${functionName}, ${message}`);
+    }
 };
 
 /**