+++ CountlyState.js
@@ -1,0 +1,6 @@
+const CountlyState = {};
+
+CountlyState.isInitialized = false
+CountlyState.CountlyReactNative = null;
+
+export default CountlyState;